feat: Implement shared Philippine time and greeting hook for consistent time display and greetings across dashboards
